requirements.txt
python-telegram-bot==20.7
youtube-transcript-api==0.6.2
anthropic==0.49.0
fpdf2==2.7.9
python-dotenv==1.0.1
yt-dlp==2024.8.6
opencv-python-headless==4.10.0.84
easyocr==1.7.1
numpy==1.26.4
python-bidi==0.4.2
